pyproject.toml
[project]
name = "project-web-cars"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"


dependencies = [
    "argon2-cffi>=25.1.0",
    "django>=5.2.7",
    "django-environ>=0.12.0",
    "django-redis>=6.0.0",
    "google-genai>=1.41.0",
    "gunicorn>=23.0.0",
    "pillow>=11.3.0",
    "psycopg>=3.2.10",
    "redis>=6.4.0",
    "whitenoise>=6.11.0",
]

[dependency-groups]
dev = [
    "django-debug-toolbar>=6.0.0",
    "django-extensions>=4.1",
    "werkzeug>=3.1.3",
]
